prefect/flows/s3-storage/s3-to-snowflake/pyproject.toml
[tool.poetry]
name = "s3-to-snowflake"
version = "0.1.0"
description = ""
authors = ["Nate Nowack <nate@prefect.io>"]

[tool.poetry.dependencies]
python = "^3.7"
snowflake-connector-python = {extras = ["pandas"], version = "^2.7.1"}
prefect = "^0.15.10"
boto3 = "^1.20.24"

[tool.poetry.dev-dependencies]
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
